Change dispatch, make result check more explicit
While this might not be 100% needed, it should make the code more
understandable. Previously, the result was evaluated to the method being
triggered. If the listener / handler does not return anything, result
might just be set to "undefined" which in term ensured that the
propagation stop would still work as intended. However, unless you know
this, it can be a bit hard to understand why that would work.

Therefore, having an explicit check for the result should make it much
more clear what the code actually does and why it works.